requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Pillow==11.1.0
requests==2.33.0
python-dotenv==1.0.1
